requirements-minimal.txt
flake8
mock
numpy
pre-commit
PyStaticConfiguration
pytest
pytest-cov
pytest-sugar
pytest-xdist
scikit-learn
scipy
tox
